feat: add tab function & install prop-types
