+++ src/component/main/Calendar.js
@@ -1,5 +1,6 @@
 import FullCalendar from "@fullcalendar/react";
-import dayGridPlugin from "@fullcalendar/daygrid"; // a plugin!
+import dayGridPlugin from "@fullcalendar/daygrid";
+
 function Calendar() {
   return <FullCalendar plugins={[dayGridPlugin]} initialView="dayGridMonth" />;
 }
